web/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
dbus-python==1.3.2
PyGObject==3.46.0
